src/api/requirements.txt
fastapi>=0.49.0
uvicorn>=0.11.7
pydantic>=1.4
pytest>=6.2.4
google-cloud-pubsub==2.13.7
grpc-google-iam-v1==0.12.4